Moved helper functions for testing to another module
